feat: Cleanup keyring implementation
