+++ xknx/secure/__init__.py
@@ -1,11 +1,9 @@
 """Classes for handling KNX IP Secure."""
 from .keyring import Keyring, load_key_ring
-from .util import decrypt_aes128cbc, extract_password, sha256_hash
+from .util import sha256_hash
 
 __all__ = [
     "Keyring",
     "load_key_ring",
     "sha256_hash",
-    "decrypt_aes128cbc",
-    "extract_password",
 ]
